data: Set DeviceIndex override for G900 wireless receiver
The default device index selected for the USB Wireless Receiver is 0xff.
This index corresponds to the unifying receivers HID++1.0 interface and
not the mouse itself. The hidpp20 driver does not handle querying or
otherwise resolving the device index for the mouse paired with the
receiver. Override the default index used for the device when matching
the G900 device ids.

In both the wireless and USB connected cases the device index is 1. The
wireless receiver is expected to handle cases where the device is off or
multiple devices are present. The hidpp20 driver does not handle the
wireless receiver to account for dynamic changes to the device presence
however if the device is not present then device index 1 will not
respond and will not be probed by the device registration. It is also
not typical for users to change the mouse(s) that are paired with the
receiver as the product is shipped with the receiver pre-paired to the
mouse.

This override was present in previous versions of the G900 device data,
however was lost between multiple changes. The wireless receiver device
id in its current form was added to the G900 device file in a04f2d1c24ab
and relies on the id being matched based on hidraw information.
Previously the device data matched against the lightspeed receivers USB
id alone, and was split from the G900 device data in c7ce1a96a7db. At
this point the device index override was present. The removal of the
lightspeed receiver device data bb3f386cca87 did not merge this override
back into the G900 device data.

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>