data: Add device data for Logitech G Pro X Wireless Superlight 2
The Logitech G Pro X Wireless Superlight 2 has a similar device data
configuration to the previous generation device (Logitech G Pro X
Wireless Superlight), using the hidpp20 driver.

The wireless receiver is capable of exposing the HID++ 2.0 interface for
the mouse, however like other Logitech wireless devices (e.g. G900) it
exposes the default index of the HID device as 0xFF for the wireless
receiver which provides an interface for querying paired devices and
there connected state. Like for other wireless receivers the hidpp20
driver does not support interfacing with the wireless receiver. To
enable access to the attached mouse the device index is overridden to
1 which provides access to the first paired/connected device on the
wireless receiver.

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>